+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+# Changelog for Unicode v1.11.0
+
+This is the changelog for Unicode v1.11.0 released on October 8th, 2020.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-unicode/unicode/tags)
+
+### Bug fixes
+
+* Rename the derived category `:visible` to `:graph` and change the definition to that in [Unicode Regular Expressions](http://unicode.org/reports/tr18/). Deprecate the derived category `:visible`.
+
 # Changelog for Unicode v1.10.0
 
 This is the changelog for Unicode v1.10.0 released on October 5th, 2020.  For older changelogs please consult the release tag on [GitHub](https://github.com/elixir-unicode/unicode/tags)
 
 ### Bug fixes
 
-* Revert "Change the definition of the derived property `All` to be the disjoint set of unicode ranges, not the closed set." since `All` in the ICU means the full range of characters.
+* Revert "Change the definition of the derived property `All` to be the disjoint set of unicode ranges, not the closed set." since `All` in the ICU means the full range of codepoints, assigned or otherwise.
 
 * Add `:inets` and `:public_key` to `:extra_applicatons` to avoid warnings on Elixir 1.11.
 
@@ -14,7 +22,6 @@ This is the changelog for Unicode v1.10.0 released on October 5th, 2020.  For ol
 
 * Rename `Unicode.ranges/0` to `Unicode.all/0` to better reflect the intent. `Unicode.ranges/0` is deprecated.
 
-### Enhancements
 
 # Changelog for Unicode v1.9.0